RR-441 - Refactored nunjucks files on Education and Training tab
