--- server/views/pages/overview/partials/educationAndTrainingOtherQualifications.njk
@@ -1,45 +1,0 @@
-<div class="govuk-grid-row">
-  <div class="govuk-grid-column-two-thirds">
-
-    <div class="govuk-summary-card">
-      <div class="govuk-summary-card__title-wrapper">
-        <h2 class="govuk-summary-card__title">Other qualifications and history</h2>
-      </div>
-      <div class="govuk-summary-card__content">
-        {% if not otherQualifications.problemRetrievingData %}
-
-          <dl class="govuk-summary-list" id="other-qualifications-list">
-            <div class="govuk-summary-list__row">
-              <dt class="govuk-summary-list__key">
-                Highest level of education
-              </dt>
-              <dd class="govuk-summary-list__value">
-                {{ otherQualifications.highestEducationLevel | formatEducationLevel }}
-              </dd>
-            </div>
-            <div class="govuk-summary-list__row">
-              <dt class="govuk-summary-list__key">
-                Other training or qualifications
-              </dt>
-              <dd class="govuk-summary-list__value">
-                <ul class="govuk-list">
-                  {% for additionalTraining in otherQualifications.additionalTraining %}
-                    <li>{{ additionalTraining | formatAdditionalTraining }}</li>
-                  {% endfor %}
-                </ul>
-              </dd>
-            </div>
-          </dl>
-
-        {% else %}
-
-          <h3 class="govuk-heading-s" data-qa="ciag-unavailable-message">We cannot show these details right now</h3>
-          <p class="govuk-body">Reload the page or try again later. Other parts of this service may still be available.</p>
-
-        {% endif %}
-
-      </div>
-    </div>
-
-  </div>
-</div>
