+++ server/views/pages/overview/partials/educationAndTrainingTab/_otherQualifications_inductionShortQuestionSet.njk
@@ -1,0 +1,16 @@
+{#
+'Other qualifications and history' summary card on the 'Education and training' tab - content for the short question set
+asked at the Induction.
+The 'Other qualifications and history' summary card is different to the one presented if the long question set was asked,
+as the follow up questions are different.
+#}
+
+<section data-qa="other-qualifications-short-question-set">
+
+  <div class="govuk-summary-card__content">
+    <dl class="govuk-summary-list" id="other-qualifications-list">
+      {# content for short question set goes here #}
+    </dl>
+  </div>
+
+</section>
