--- server/views/pages/overview/partials/educationAndTrainingTabContents.njk
@@ -1,3 +1,0 @@
-{% include './educationAndTrainingFunctionalSkills.njk' %}
-{% include './educationAndTrainingInPrisonQualifications.njk' %}
-{% include './educationAndTrainingOtherQualifications.njk' %}
