+++ server/views/pages/overview/partials/overviewTab/functionalSkillsSidebar.njk
